requirements.txt
llama-cpp-python==0.2.24
pandas
streamlit
fastapi
uvicorn
langchain
openai
python-dotenv
tqdm
torch
openai  
langchain
openpyxl    
transformers
ctransformers
tabulate
fuzzywuzzy
pydantic
langsmith
langchain_openai
langgraph
python-Levenshtein   

